Change Server to output JSON formatted responses
As I misunderstood the original requirements of the project, I had developed
it to output HTML to the browser, now looking back and actually searching what
API means I understand this is not the case. This revision is more for my
benefit than an attempt to fix what I had made for the project, and as such
is in a different tree.